function/requirements.txt
annotated-types==0.6.0
asn1crypto==1.5.1
attrs==23.1.0
aws-sam-translator==1.81.0
aws-xray-sdk==2.12.1
beautifulsoup4==4.12.2
blinker==1.7.0
boto3==1.29.5
botocore==1.32.5
certifi==2023.11.17
cffi==1.16.0
cfn-lint==0.83.3
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
docker==6.1.3
ecdsa==0.18.0
Flask==3.0.0
Flask-Cors==4.0.0
graphql-core==3.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jschema-to-python==1.2.3
jsondiff==2.0.0
jsonpatch==1.33
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-path==0.3.2
jsonschema-specifications==2023.11.1
junit-xml==1.9
lazy-object-proxy==1.9.0
lxml==4.9.3
MarkupSafe==2.1.3
moto==4.2.10
mpmath==1.3.0
multipart==0.2.4
networkx==3.2.1
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
packaging==23.2
pathable==0.4.3
pbr==6.0.0
psycopg2-binary==2.9.9
py-partiql-parser==0.4.2
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
python-jose==3.3.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
responses==0.24.1
rfc3339-validator==0.1.4
rpds-py==0.13.1
rsa==4.9
s3transfer==0.7.0
sarif-om==1.0.4
scramp==1.4.4
six==1.16.0
soupsieve==2.5
sshpubkeys==3.3.1
sympy==1.12
typing_extensions==4.8.0
urllib3==2.0.7
websocket-client==1.6.4
Werkzeug==3.0.1
wrapt==1.16.0
xmltodict==0.13.0
